refactor(cli): split component into api/component leaf shape
Reorganize the flat `api/component/component.mjs` into the fractal leaf
shape over a shared `_adapter.mjs` resolver:

- `component.mjs` becomes a dispatcher + barrel: it parses (name, options),
  resolves the subject once via `_adapter`, routes to the correct leaf, and
  re-exports every leaf for direct scripting use.
- `_adapter.mjs` holds ALL resolution + I/O, deduped: core/external/scoped/
  integration ownership, ambiguity handling, fuzzy search, doc loading, prop
  extraction, and ownership shaping.
- Leaves are thin projections into one envelope each: `list` (component.list
  with data.detail names|compact|full), `detail`, `detail/props`,
  `detail/source`, `detail/showcase`, `detail/blocks`.

Pure reorg: `--json` and human output are byte-identical across all modes.
Types stay central in `types/component.d.ts`; all existing tests pass.
