+++ packages/cli/api/component/component.mjs
@@ -1,91 +1,47 @@
 // Copyright (c) Meta Platforms, Inc. and affiliates.
 
 /**
- * @file Programmatic API for the component command.
+ * @file Programmatic API for the component command — DISPATCHER + BARREL.
  *
- * Returns the same typed envelope { type, data } that `astryx --json component` outputs.
- * The CLI command handler is a thin wrapper around this function.
+ * Returns the same typed envelope { type, data } that `astryx --json component`
+ * outputs. `component(name, opts)` parses the (name, options) pair, resolves the
+ * subject once via `_adapter`, and routes to the correct leaf
+ * (list · detail · detail.props/source/showcase/blocks), returning that leaf's
+ * envelope. The CLI command handler is a thin wrapper around this function.
+ *
+ * Every leaf is also re-exported for direct scripting use.
  */
 
-import * as fs from 'node:fs';
-import * as path from 'node:path';
 import {ERROR_CODES} from '../../lib/error-codes.mjs';
-import {findCoreDir, discoverExternalPackages} from '../../utils/paths.mjs';
+import {AstryxError} from '../error.mjs';
 import {
   CORE_PACKAGE,
-  discoverComponents,
-  discoverExternalComponentsGrouped,
-  discoverIntegrationComponents,
-  findComponentReadme,
-  findComponentSource,
-  findExternalComponentDoc,
-  findIntegrationComponentDoc,
-  findIntegrationComponentSource,
-  resolveImportPath,
-} from '../../lib/component-discovery.mjs';
-import {Project} from '../../lib/project.mjs';
-import {loadDocs} from '../../lib/component-loader.mjs';
-import {searchComponents} from '../../lib/string-utils.mjs';
-import {AstryxError} from '../error.mjs';
-import {findShowcase, findRelatedBlocks} from '../template/template.mjs';
+  requireCoreDir,
+  loadIntegrationsSafely,
+  resolveOwners,
+  classifyScope,
+  assertUnambiguousOwners,
+  resolveLegacyExternalDoc,
+  resolveCoreSourcePath,
+  resolveUnscopedDoc,
+  loadComponentDoc,
+  scopeSubComponent,
+} from './_adapter.mjs';
+import {componentList} from './list/list.mjs';
+import {componentDetail} from './detail/detail.mjs';
+import {componentDetailProps} from './detail/props/props.mjs';
+import {componentDetailSource} from './detail/source/source.mjs';
+import {componentDetailShowcase} from './detail/showcase/showcase.mjs';
+import {componentDetailBlocks} from './detail/blocks/blocks.mjs';
 
-/**
- * A loaded component doc. `loadDocs` returns the authored `.doc.mjs` shape,
- * which is either a single-component or multi-component doc; this loose view
- * captures the fields the API reads across both forms.
- * @typedef {object} LoadedComponentDoc
- * @property {string} [name]
- * @property {string} [description]
- * @property {any[]} [props]
- * @property {any[]} [components]
- * @property {{description?: string}} [usage]
- * @property {any} [theming]
- */
-
-/**
- * Options object for `loadDocs`, matching its declared parameter shape (used
- * as a cast target so `lang` (which the API may hold as `string|null`) type
- * checks against `loadDocs`'s `lang?: string`).
- * @typedef {{zh?: boolean, dense?: boolean, lang?: string}} LoadDocsOpts
- */
-
-/**
- * Load the configured integrations for `cwd`, swallowing any config errors so
- * component discovery never hard-fails on a malformed/absent integration. An
- * empty list means "core only".
- * @param {string} cwd
- * @returns {Promise<import('../../lib/integrations.mjs').LoadedIntegration[]>}
- */
-async function loadIntegrationsSafely(cwd) {
-  try {
-    const project = await Project.load(cwd);
-    return /** @type {import('../../lib/integrations.mjs').LoadedIntegration[]} */ (
-      project.loadedIntegrations
-    );
-  } catch {
-    return [];
-  }
-}
-
-/**
- * Resolve a loaded integration by package name.
- * @param {import('../../lib/integrations.mjs').LoadedIntegration[]} loadedIntegrations
- * @param {string} packageName
- */
-function findLoadedIntegration(loadedIntegrations, packageName) {
-  return loadedIntegrations.find(i => i.name === packageName) ?? null;
-}
-
-/**
- * Resolve an external package by name from the discovered externals list.
- * @param {string} packageName - e.g. '@acme/xds-widgets'
- * @param {string} cwd
- * @returns {{ name: string, category: string, docsDir: string } | null}
- */
-function resolveExternalPackage(packageName, cwd) {
-  const externals = discoverExternalPackages(cwd);
-  return externals.find(ext => ext.name === packageName) ?? null;
-}
+export {
+  componentList,
+  componentDetail,
+  componentDetailProps,
+  componentDetailSource,
+  componentDetailShowcase,
+  componentDetailBlocks,
+};
 
 /**
  * @param {string} [name]
@@ -102,7 +58,14 @@ function resolveExternalPackage(packageName, cwd) {
  * @param {string} [options.lang]
  * @param {boolean} [options.zh]
  * @param {boolean} [options.dense]
- * @returns {Promise<{type: string, data: unknown}>}
+ * @returns {Promise<(
+ *   import('../../types/component').ComponentListResponse
+ *   | import('../../types/component').ComponentDetailResponse
+ *   | import('../../types/component').ComponentDetailPropsResponse
+ *   | import('../../types/component').ComponentDetailSourceResponse
+ *   | import('../../types/component').ComponentDetailShowcaseResponse
+ *   | import('../../types/component').ComponentDetailBlocksResponse
+ * )>}
  */
 export async function component(name, options = {}) {
   const {
@@ -127,183 +90,14 @@ export async function component(name, options = {}) {
   const isListView = list || category != null || !name;
   const detail = detailOption ?? (isListView ? 'brief' : 'full');
 
-  const coreDir = findCoreDir(cwd);
-  if (!coreDir) {
-    throw new AstryxError('Could not find @astryxdesign/core package', undefined, ERROR_CODES.ERR_CORE_NOT_FOUND);
-  }
+  const coreDir = requireCoreDir(cwd);
 
   // ── List mode ──────────────────────────────────────────────────
-
   if (category || list || !name) {
-    const components = discoverComponents(coreDir);
-
-    if (category) {
-      const match = Object.entries(components).find(
-        ([key]) => key.toLowerCase() === category.toLowerCase(),
-      );
-      if (!match) {
-        throw new AstryxError(
-          `Unknown category "${category}"`,
-          Object.keys(components).map(k => ({name: k, reason: 'valid category'})),
-          ERROR_CODES.ERR_UNKNOWN_CATEGORY,
-        );
-      }
-
-      if (detail === 'compact') {
-        const entries = [];
-        for (const comp of match[1]) {
-          const readme = findComponentReadme(coreDir, comp);
-          if (readme && readme.endsWith('.doc.mjs')) {
-            try {
-              const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(readme, /** @type {LoadDocsOpts} */ ({zh, lang})));
-              entries.push({name: comp, description: docs.usage?.description || docs.description || '', import: resolveImportPath(coreDir, comp)});
-            } catch {
-              entries.push({name: comp, description: '', import: resolveImportPath(coreDir, comp)});
-            }
-          } else {
-            entries.push({name: comp, description: '', import: resolveImportPath(coreDir, comp)});
-          }
-        }
-        return {type: 'component.list', data: {detail: 'compact', components: {[match[0]]: entries}}};
-      }
-
-      if (detail === 'full') {
-        const entries = [];
-        for (const comp of match[1]) {
-          const readme = findComponentReadme(coreDir, comp);
-          if (readme && readme.endsWith('.doc.mjs')) {
-            try {
-              entries.push(await loadDocs(readme, /** @type {LoadDocsOpts} */ ({zh, lang, dense})));
-            } catch {
-              entries.push({name: `XDS${comp}`, description: ''});
-            }
-          } else {
-            entries.push({name: `XDS${comp}`, description: ''});
-          }
-        }
-        return {type: 'component.list', data: {detail: 'full', components: {[match[0]]: entries}}};
-      }
-
-      // Default: brief — package-qualified object list for the category.
-      // Pre-1.0 JSON contract: members are {name, package} objects, not bare
-      // strings, so consumers can disambiguate ownership.
-      return {
-        type: 'component.list',
-        data: {detail: 'names', components: {[match[0]]: match[1].map(n => ({name: n, package: CORE_PACKAGE}))}},
-      };
-    }
-
-    // All components — merge core + external packages with grouped subcategories
-    if (detail === 'compact') {
-      /** @type {Record<string, Array<import('../../types/component').ComponentBriefEntry>>} */
-      const result = {};
-      for (const [cat, comps] of Object.entries(components)) {
-        result[cat] = [];
-        for (const comp of comps) {
-          const readme = findComponentReadme(coreDir, comp);
-          if (readme && readme.endsWith('.doc.mjs')) {
-            try {
-              const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(readme, /** @type {LoadDocsOpts} */ ({zh, lang})));
-              result[cat].push({name: comp, description: docs.usage?.description || docs.description || '', import: resolveImportPath(coreDir, comp)});
-            } catch {
-              result[cat].push({name: comp, description: '', import: resolveImportPath(coreDir, comp)});
-            }
-          } else {
-            result[cat].push({name: comp, description: '', import: resolveImportPath(coreDir, comp)});
-          }
-        }
-      }
-      return {type: 'component.list', data: {detail: 'compact', components: result}};
-    }
-
-    if (detail === 'full') {
-      /** @type {Record<string, Array<unknown>>} */
-      const result = {};
-      for (const [cat, comps] of Object.entries(components)) {
-        result[cat] = [];
-        for (const comp of comps) {
-          const readme = findComponentReadme(coreDir, comp);
-          if (readme && readme.endsWith('.doc.mjs')) {
-            try {
-              result[cat].push(await loadDocs(readme, /** @type {LoadDocsOpts} */ ({zh, lang, dense})));
-            } catch {
-              result[cat].push({name: `XDS${comp}`, description: ''});
-            }
-          } else {
-            result[cat].push({name: `XDS${comp}`, description: ''});
-          }
-        }
-      }
-      return {type: 'component.list', data: {detail: 'full', components: result}};
-    }
-
-    // Default: brief — package-qualified object list (core + integrations).
-    // Pre-1.0 JSON contract: each group's members are {name, package} objects.
-    /** @type {Record<string, Array<{name: string, package: string}>>} */
-    const listData = {};
-    for (const [cat, comps] of Object.entries(components)) {
-      listData[cat] = comps.map(n => ({name: n, package: CORE_PACKAGE}));
-    }
-
-    // Integration components (authoritative source: loadedIntegrations).
-    const loadedIntegrations = await loadIntegrationsSafely(cwd);
-    const seenIntegration = new Set();
-    for (const integration of loadedIntegrations) {
-      seenIntegration.add(integration.name);
-      const owned = discoverIntegrationComponents(integration);
-      // Group integration components by their doc `group`, falling back to the
-      // package name. Keys are package-qualified so they never collide with
-      // core groups or each other.
-      /** @type {Map<string, Array<{name: string, package: string}>>} */
-      const byGroup = new Map();
-      for (const rec of owned) {
-        const groupLabel = rec.group ?? integration.name;
-        const key = `${groupLabel} (${integration.name})`;
-        if (!byGroup.has(key)) byGroup.set(key, []);
-        byGroup.get(key)?.push({name: rec.name, package: integration.name});
-      }
-      for (const [key, members] of byGroup) {
-        members.sort((a, b) => a.name.localeCompare(b.name));
-        listData[key] = members;
-      }
-    }
-
-    // Back-compat: node_modules-scanned external packages (pkg.astryx.docs)
-    // that are NOT configured integrations. Preserves existing discovery for
-    // consumers that haven't adopted the config-integration flow.
-    const externals = discoverExternalPackages(cwd);
-    for (const ext of externals) {
-      if (seenIntegration.has(ext.name)) continue;
-      const grouped = discoverExternalComponentsGrouped(ext.docsDir);
-      const groupKeys = Object.keys(grouped);
-      if (groupKeys.length === 0) continue;
-
-      const hasGroups = groupKeys.some(
-        k => grouped[k].length > 1 || grouped[k][0] !== k,
-      );
-
-      if (hasGroups) {
-        for (const [group, members] of Object.entries(grouped)) {
-          listData[`${group} (${ext.name})`] = members.map(n => ({
-            name: n,
-            package: ext.name,
-          }));
-        }
-      } else {
-        const allComps = Object.values(grouped).flat().sort();
-        if (allComps.length > 0) {
-          listData[`${ext.category} (${ext.name})`] = allComps.map(n => ({
-            name: n,
-            package: ext.name,
-          }));
-        }
-      }
-    }
-    return {type: 'component.list', data: {detail: 'names', components: listData}};
+    return componentList(coreDir, {cwd, category, detail, zh, dense, lang});
   }
 
   // ── Single component ───────────────────────────────────────────
-
   if (typeof name !== 'string') {
     throw new AstryxError(
       `No component named "${String(name)}"`,
@@ -313,347 +107,100 @@ export async function component(name, options = {}) {
   }
 
   const dirName = name.replace(/^XDS/, '');
+  const docOpts = {zh, dense, lang};
 
-  // Ownership-aware resolution. Build the set of OWNER packages that provide a
-  // component with this name across core + every loaded integration. This is
-  // what lets the CLI disambiguate by package and expose the owner's source +
-  // issuesUrl (the inputs the future integration-component swizzle needs).
+  // Ownership-aware resolution: who provides `dirName` across core + every
+  // loaded integration. The scoped/ambiguity/fuzzy resolution below all read
+  // this set (built once, in the adapter).
   const loadedIntegrations = await loadIntegrationsSafely(cwd);
-  const coreDocPath = findComponentReadme(coreDir, dirName);
-  /**
-   * @type {Array<{
-   *   package: string,
-   *   docPath: string,
-   *   sourcePath: string|null,
-   *   issuesUrl: string|undefined,
-   *   integration: object|null,
-   * }>}
-   */
-  const owners = [];
-  if (coreDocPath) {
-    owners.push({
-      package: CORE_PACKAGE,
-      docPath: coreDocPath,
-      sourcePath: findComponentSource(coreDir, dirName),
-      issuesUrl: undefined,
-      integration: null,
-    });
-  }
-  for (const integration of loadedIntegrations) {
-    const docPath = findIntegrationComponentDoc(integration, dirName);
-    if (!docPath) continue;
-    owners.push({
-      package: integration.name,
-      docPath,
-      sourcePath: findIntegrationComponentSource(integration, dirName),
-      issuesUrl: integration.issuesUrl,
-      integration,
-    });
-  }
+  const owners = resolveOwners(coreDir, dirName, loadedIntegrations);
 
-  /**
-   * Augment a loaded `component.detail` doc with ownership metadata. Adds
-   * `package`, the resolved `import` specifier, and `sourceAvailable` (whether
-   * a swizzleable source file exists for the owner). Existing doc fields
-   * (name, usage, props, …) are preserved.
-   * @param {LoadedComponentDoc} docs
-   * @param {{package: string, sourcePath: string|null}} owner
-   * @param {string} componentName
-   */
-  function withOwnership(docs, owner, componentName) {
-    const importSpec =
-      owner.package === CORE_PACKAGE
-        ? resolveImportPath(/** @type {string} */ (coreDir), componentName)
-        : `${owner.package}/${componentName}`;
-    return {
-      ...docs,
-      package: owner.package,
-      import: importSpec,
-      sourceAvailable: owner.sourcePath != null,
-    };
-  }
-
-  // When scoped to a specific package, search that package first.
-  // This is critical for components that exist in both core and an external
-  // package (e.g. AppShell, Button, SideNav) — the package scope ensures
-  // the external package's docs are returned, not core's.
+  // ── Scoped to a specific package (--package) ───────────────────
+  // Searches that package first — critical for names that exist in both core
+  // and an external package (AppShell, Button, SideNav).
   if (packageScope) {
-    // Core scope: resolve from core directly.
-    if (packageScope === CORE_PACKAGE) {
-      const owner = owners.find(o => o.package === CORE_PACKAGE);
-      if (!owner) {
-        throw new AstryxError(`No component "${name}" in package "${packageScope}"`, undefined, ERROR_CODES.ERR_UNKNOWN_COMPONENT);
-      }
+    const scoped = classifyScope(packageScope, {owners, loadedIntegrations, cwd, name});
+
+    if (scoped.kind === 'core' || scoped.kind === 'integration') {
+      const owner = scoped.owner;
       if (source) {
-        if (!owner.sourcePath) {
-          throw new AstryxError(`Source for "${name}" not found`, undefined, ERROR_CODES.ERR_NO_SOURCE);
-        }
-        return {type: 'component.detail.source', data: {component: dirName, source: fs.readFileSync(owner.sourcePath, 'utf-8')}};
+        return componentDetailSource(dirName, owner.sourcePath, {
+          name,
+          notFoundInPackage: scoped.kind === 'integration' ? packageScope : null,
+        });
       }
-      const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(owner.docPath, /** @type {LoadDocsOpts} */ ({zh, dense, lang})));
-      if (props) {
-        const p = docs.props || (docs.components ? docs.components.flatMap(c => c.props || []) : []);
-        return {type: 'component.detail.props', data: p};
-      }
-      return {type: 'component.detail', data: withOwnership(docs, owner, dirName)};
+      const docs = await loadComponentDoc(owner.docPath, docOpts);
+      if (props) return componentDetailProps(docs);
+      return componentDetail(docs, owner, dirName, coreDir);
     }
 
-    // Integration scope (authoritative): resolve from the loaded integration.
-    const integration = findLoadedIntegration(loadedIntegrations, packageScope);
-    if (integration) {
-      const owner = owners.find(o => o.package === packageScope);
-      if (!owner) {
-        throw new AstryxError(`No component "${name}" in package "${packageScope}"`, undefined, ERROR_CODES.ERR_UNKNOWN_COMPONENT);
-      }
-      if (source) {
-        if (!owner.sourcePath) {
-          throw new AstryxError(`Source for "${name}" not found in package "${packageScope}"`, undefined, ERROR_CODES.ERR_NO_SOURCE);
-        }
-        return {type: 'component.detail.source', data: {component: dirName, source: fs.readFileSync(owner.sourcePath, 'utf-8')}};
-      }
-      const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(owner.docPath, /** @type {LoadDocsOpts} */ ({zh, dense, lang})));
-      if (props) {
-        const p = docs.props || (docs.components ? docs.components.flatMap(c => c.props || []) : []);
-        return {type: 'component.detail.props', data: p};
-      }
-      return {type: 'component.detail', data: withOwnership(docs, owner, dirName)};
-    }
-
-    // Legacy fallback: node_modules `pkg.astryx.docs` external package.
-    const ext = resolveExternalPackage(packageScope, cwd);
-    if (!ext) {
-      throw new AstryxError(`External package "${packageScope}" not found`, undefined, ERROR_CODES.ERR_UNKNOWN_PACKAGE);
-    }
-
+    // Legacy `pkg.astryx.docs` external package.
     if (showcase) {
-      const match = await findShowcase(dirName, cwd, {package: packageScope});
-      if (!match) {
-        throw new AstryxError(`No showcase found for "${name}" in package "${packageScope}"`, undefined, ERROR_CODES.ERR_NO_SHOWCASE);
-      }
-      return {
-        type: 'component.detail.showcase',
-        data: {
-          component: dirName,
-          aspectRatio: match.aspectRatio,
-          source: fs.readFileSync(match.filePath, 'utf-8'),
-        },
-      };
+      return componentDetailShowcase(dirName, {cwd, name, packageScope});
     }
-
-    const extDocPath = findExternalComponentDoc(ext.docsDir, dirName);
-    if (extDocPath && extDocPath.endsWith('.doc.mjs')) {
-      const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(extDocPath, /** @type {LoadDocsOpts} */ ({zh, dense, lang})));
-
-      if (props) {
-        const p = docs.props || (docs.components ? docs.components.flatMap(c => c.props || []) : []);
-        return {type: 'component.detail.props', data: p};
-      }
-      return {
-        type: 'component.detail',
-        data: withOwnership(docs, {package: ext.name, sourcePath: null}, dirName),
-      };
+    const extDocPath = resolveLegacyExternalDoc(scoped.ext, dirName);
+    if (extDocPath) {
+      const docs = await loadComponentDoc(extDocPath, docOpts);
+      if (props) return componentDetailProps(docs);
+      return componentDetail(docs, {package: scoped.ext.name, sourcePath: null}, dirName, coreDir);
     }
     throw new AstryxError(`No component "${name}" in package "${packageScope}"`, undefined, ERROR_CODES.ERR_UNKNOWN_COMPONENT);
   }
 
-  // Ambiguity: when the name is owned by MORE THAN ONE package (core and/or
-  // integrations) and the caller did not scope with --package, refuse to guess.
-  // NOTE: legacy `pkg.astryx.docs` externals are intentionally NOT part of this
-  // ambiguity set — they retain their historical core-first fallback below so
-  // existing consumers (and tests) keep working. Only config-driven integration
-  // ownership participates here.
-  if (owners.length > 1) {
-    throw new AstryxError(
-      `Component "${dirName}" is provided by multiple packages. Re-run with --package <pkg> to choose one.`,
-      owners.map(o => ({name: o.package, reason: 'provides this component'})),
-      ERROR_CODES.ERR_UNKNOWN_COMPONENT,
-    );
-  }
+  // ── Ambiguity: owned by MORE THAN ONE package, no --package ─────
+  assertUnambiguousOwners(owners, dirName);
 
-  // Single non-core owner (an integration provides it, core does not) — resolve
-  // from that integration so the integration component is authoritative.
+  // ── Single non-core owner (an integration provides it, core does not) ──
   if (owners.length === 1 && owners[0].package !== CORE_PACKAGE) {
     const owner = owners[0];
     if (source) {
-      if (!owner.sourcePath) {
-        throw new AstryxError(`Source for "${name}" not found`, undefined, ERROR_CODES.ERR_NO_SOURCE);
-      }
-      return {type: 'component.detail.source', data: {component: dirName, source: fs.readFileSync(owner.sourcePath, 'utf-8')}};
+      return componentDetailSource(dirName, owner.sourcePath, {name});
     }
     if (showcase) {
-      throw new AstryxError(`No showcase found for "${name}"`, undefined, ERROR_CODES.ERR_NO_SHOWCASE);
+      // Integration components don't carry showcases — fail without scanning.
+      return componentDetailShowcase(dirName, {cwd, name, resolve: false});
     }
-    const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(owner.docPath, /** @type {LoadDocsOpts} */ ({zh, dense, lang})));
-    if (props) {
-      const p = docs.props || (docs.components ? docs.components.flatMap(c => c.props || []) : []);
-      return {type: 'component.detail.props', data: p};
-    }
-    return {type: 'component.detail', data: withOwnership(docs, owner, dirName)};
+    const docs = await loadComponentDoc(owner.docPath, docOpts);
+    if (props) return componentDetailProps(docs);
+    return componentDetail(docs, owner, dirName, coreDir);
   }
 
+  // ── No-scope core path ─────────────────────────────────────────
+  // `--source` reads core directly (no external/fuzzy fallback).
   if (source) {
-    const sourcePath = findComponentSource(coreDir, dirName);
-    if (!sourcePath) {
-      throw new AstryxError(`Source for "${name}" not found`, undefined, ERROR_CODES.ERR_NO_SOURCE);
-    }
-    return {type: 'component.detail.source', data: {component: dirName, source: fs.readFileSync(sourcePath, 'utf-8')}};
+    return componentDetailSource(dirName, resolveCoreSourcePath(coreDir, dirName), {name});
   }
-
   if (showcase) {
-    const match = await findShowcase(dirName, cwd);
-    if (!match) {
-      throw new AstryxError(`No showcase found for "${name}"`, undefined, ERROR_CODES.ERR_NO_SHOWCASE);
-    }
-    return {
-      type: 'component.detail.showcase',
-      data: {
-        component: dirName,
-        aspectRatio: match.aspectRatio,
-        source: fs.readFileSync(match.filePath, 'utf-8'),
-      },
-    };
+    return componentDetailShowcase(dirName, {cwd, name});
   }
 
-  let readmePath = findComponentReadme(coreDir, dirName);
-  let resolvedName = dirName;
-  // Track the resolving owner so the detail payload can carry ownership info.
-  // Defaults to core; the legacy-external fallback below may reassign it.
-  let resolvedOwnerPackage = CORE_PACKAGE;
-  let resolvedSourcePath = readmePath ? findComponentSource(coreDir, dirName) : null;
-
-  if (!readmePath) {
-    const externals = discoverExternalPackages(cwd);
-    for (const ext of externals) {
-      const extDocPath = findExternalComponentDoc(ext.docsDir, dirName);
-      if (extDocPath) {
-        readmePath = extDocPath;
-        resolvedOwnerPackage = ext.name;
-        resolvedSourcePath = null;
-        break;
-      }
-    }
-  }
-
-  if (!readmePath) {
-    const components = discoverComponents(coreDir);
-    const results = await searchComponents(dirName, coreDir, components);
-
-    if (results.length > 0) {
-      const topScore = results[0].score;
-      const topTied = results.filter(r => r.score === topScore);
-      const secondScore = results.length > topTied.length ? results[topTied.length].score : 0;
-      const gap = topScore - secondScore;
-
-      if (topScore >= 90 && topTied.length === 1 && gap >= 20) {
-        resolvedName = topTied[0].name;
-        readmePath = findComponentReadme(coreDir, resolvedName);
-        resolvedOwnerPackage = CORE_PACKAGE;
-        resolvedSourcePath = findComponentSource(coreDir, resolvedName);
-      } else {
-        const threshold = Math.max(topScore - 20, 1);
-        const candidates = results.filter(r => r.score >= threshold).slice(0, 5);
-        if (candidates.length < 2) candidates.push(...results.slice(candidates.length, 2));
-        throw new AstryxError(
-          `No component named "${name}"`,
-          candidates.map(c => ({name: c.name, reason: c.reason})),
-          ERROR_CODES.ERR_UNKNOWN_COMPONENT,
-        );
-      }
-    } else {
-      throw new AstryxError(`No component named "${name}"`, undefined, ERROR_CODES.ERR_UNKNOWN_COMPONENT);
-    }
-  }
-
-  if (!readmePath || !readmePath.endsWith('.doc.mjs')) {
-    throw new AstryxError(`No .doc.mjs found for "${resolvedName}". The component needs a typed doc file.`, undefined, ERROR_CODES.ERR_NO_DOC);
-  }
-
-  const docs = /** @type {LoadedComponentDoc} */ (await loadDocs(readmePath, /** @type {LoadDocsOpts} */ ({zh, dense, lang})));
+  const resolved = await resolveUnscopedDoc(dirName, {coreDir, cwd, name});
+  const docs = await loadComponentDoc(resolved.readmePath, docOpts);
 
   // ── Blocks mode ──────────────────────────────────────────────
   if (blocks) {
-    const allBlocks = await findRelatedBlocks(dirName);
-    const toEntry = (/** @type {any} */ b) => ({
-      name: b.dirName,
-      displayName: b.name,
-      description: b.description,
-      isShowcase: b.isShowcase ?? false,
-      category: b.category,
-    });
-
-    // Examples: blocks in the component's own directory, or
-    // componentsUsed match for sub-components without a directory.
-    const ownDir = allBlocks.filter((/** @type {any} */ b) => path.basename(b.category) === dirName);
-    const examples = ownDir.length > 0
-      ? ownDir
-      : allBlocks.filter(b => b.componentsUsed?.some(c => c === dirName));
-    const exampleSet = new Set(examples.map(b => b.dirName));
-
-    // Showcase: the single hero example from the examples list.
-    const showcaseBlock = examples.find(b => b.isShowcase) || null;
-
-    // Related: everything else that uses this component but isn't
-    // primarily about it (e.g. a Dialog block that has a Button).
-    const related = allBlocks.filter(b => !exampleSet.has(b.dirName));
-
-    return {
-      type: 'component.detail.blocks',
-      data: {
-        component: dirName,
-        showcase: showcaseBlock ? toEntry(showcaseBlock) : null,
-        examples: examples.filter(b => b !== showcaseBlock).map(toEntry),
-        related: related.map(toEntry),
-      },
-    };
+    return componentDetailBlocks(dirName);
   }
 
   // ── Sub-component scoping ────────────────────────────────────
   // If the user asked for "Code" but the doc is for "CodeBlock" (parent),
   // scope the response to just the matching sub-component.
-  const requestedXDS = `XDS${dirName}`;
-  const isParentDoc = docs.name && docs.name.toLowerCase() !== dirName.toLowerCase();
-  const matchingComponent = isParentDoc && docs.components
-    ? docs.components.find(c => c.name === requestedXDS || c.name === dirName)
-    : null;
-
-  if (matchingComponent) {
-    /** @type {LoadedComponentDoc & {parentDoc?: string, import?: string}} */
-    const scoped = {
-      name: dirName,
-      description: matchingComponent.description,
-      props: matchingComponent.props,
-      // Keep parent context for reference
-      components: [matchingComponent],
-      parentDoc: docs.name,
-      import: resolveImportPath(coreDir, dirName),
-    };
-    if (docs.usage) scoped.usage = docs.usage;
-    if (docs.theming) scoped.theming = docs.theming;
-
-    if (props) {
-      return {type: 'component.detail.props', data: matchingComponent.props || []};
-    }
-    return {
-      type: 'component.detail',
-      data: withOwnership(
-        scoped,
-        {package: resolvedOwnerPackage, sourcePath: resolvedSourcePath},
-        dirName,
-      ),
-    };
+  const sub = scopeSubComponent(docs, dirName, coreDir);
+  if (sub) {
+    if (props) return componentDetailProps({props: sub.matchingComponent.props});
+    return componentDetail(
+      sub.scoped,
+      {package: resolved.resolvedOwnerPackage, sourcePath: resolved.resolvedSourcePath},
+      dirName,
+      coreDir,
+    );
   }
 
-  if (props) {
-    const p = docs.props || (docs.components ? docs.components.flatMap(c => c.props || []) : []);
-    return {type: 'component.detail.props', data: p};
-  }
-
-  return {
-    type: 'component.detail',
-    data: withOwnership(
-      docs,
-      {package: resolvedOwnerPackage, sourcePath: resolvedSourcePath},
-      resolvedName,
-    ),
-  };
+  if (props) return componentDetailProps(docs);
+  return componentDetail(
+    docs,
+    {package: resolved.resolvedOwnerPackage, sourcePath: resolved.resolvedSourcePath},
+    resolved.resolvedName,
+    coreDir,
+  );
 }
